pyproject.toml
[project]
name = "CondaRTS"
version = "0.0.3"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "pygame-ce>=2.5.5",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "ruff>=0.13.1",
]

[tool.ruff.lint]
select = [
    "ANN",  # flake8-annotations
    "B",  # flake8-bugbear
    "F",  # Pyflakes
    "FBT",  # flake8-boolean-trap
    "I",  # isort
    "TC",  # flake8-type-checking
    "RET",  # flake8-return
    "UP",  # pyupgrade
]
ignore = [
    "ANN002",  # Missing type annotation for *args
    "ANN003",  # Missing type annotation for **kwargs
    "B905",  # `zip()` without an explicit `strict=` parameter
]
extend-safe-fixes = [
    "TC",  # flake8-type-checking
]

[tool.ruff.lint.per-file-ignores]
"CondaRTS.py" = [
    "B023",  # Function definition does not bind loop variable
    "COM812",  # Trailing comma missing - conflicts with ruff format
]
